refactoring FaceDetection package
