+++ FaceDetection/faceDetector.py
@@ -1,20 +1,17 @@
-# from faceClassifer import *
 import os
 import pickle
 from collections import defaultdict
-import PIL
 import mmcv
-import seaborn as sns
-import pandas as pd
 import tqdm
 from PIL import Image
-from torchvision import transforms
 from DataProcessing.DB.dal import get_entries, Crop
 from DataProcessing.dataProcessingConstants import ID_TO_NAME, NAME_TO_ID
+from FaceDetection.augmentions import normalize_image, crop_top_third_and_sides
+from FaceDetection.data_handler import load_data, is_img
 from FaceDetection.facenet_pytorch import MTCNN
 import numpy as np
 from matplotlib import pyplot as plt
-import torch
+
 
 class FaceDetector():
     """
@@ -25,40 +22,39 @@ class FaceDetector():
     the correct format.
     """
     def __init__(self, faces_data_path:str=None, thresholds=[0.8,0.8,0.8],
-                 keep_all=False, device='cuda:1'):
+                 keep_all=False, device='cuda:1',min_face_size=20):
         self.device = device
         self.faces_data_path = faces_data_path
         self.keep_all = keep_all
         self.facenet_detecor = MTCNN(margin=40, select_largest=True, post_process=False, device=device,
-                                     keep_all=self.keep_all, thresholds=thresholds)
-        self.high_conf_face_imgs = defaultdict(list)
+                                     keep_all=self.keep_all, thresholds=thresholds, min_face_size=min_face_size)
 
-    def crop_top_third_and_sides(self, img, is_PIL_input):
-        # this assumes img is a PIL obj
-        if not is_PIL_input:
-            img = PIL.Image.fromarray(img)
-        width, height = img.size
-        cropped_img = img.crop((width*0.2, 0, width*0.8, height*0.3))
-        return cropped_img
+    # def crop_top_third_and_sides(self, img, is_PIL_input):
+    #     # this assumes img is a PIL obj
+    #     if not is_PIL_input:
+    #         img = PIL.Image.fromarray(img)
+    #     width, height = img.size
+    #     cropped_img = img.crop((width*0.2, 0, width*0.8, height*0.3))
+    #     return cropped_img
 
-    def normalize_image(self, img):
+    # def normalize_image(self, img):
+    #
+    #     transform = transforms.Compose([
+    #         transforms.Normalize((0.485, 0.456, 0.406), (0.229, 0.224, 0.225)),
+    #     ])
+    #     return transform(img)
 
-        transform = transforms.Compose([
-            transforms.Normalize((0.485, 0.456, 0.406), (0.229, 0.224, 0.225)),
-        ])
-        return transform(img)
-
-    def detect_single_face_inv_norm(self, img):
-        # use face detector to find a single face in an image, rests to entered init of keeping face after change
-        self.facenet_detecor.keep_all = False
-        comp = transforms.Compose([
-            transforms.ToTensor(),
-            lambda x:x*255,
-        ])
-        new_im = comp(img).permute(1,2,0).int()
-        ret, prob = self.facenet_detecor(new_im, return_prob=True)
-        self.facenet_detecor.keep_all = self.keep_all
-        return ret , prob
+    # def detect_single_face_inv_norm(self, img):
+    #     # use face detector to find a single face in an image, rests to entered init of keeping face after change
+    #     self.facenet_detecor.keep_all = False
+    #     comp = transforms.Compose([
+    #         transforms.ToTensor(),
+    #         lambda x:x*255,
+    #     ])
+    #     new_im = comp(img).permute(1,2,0).int()
+    #     ret, prob = self.facenet_detecor(new_im, return_prob=True)
+    #     self.facenet_detecor.keep_all = self.keep_all
+    #     return ret , prob
 
     def detect_single_face(self, img):
         # use face detector to find a single face in an image, rests to entered init of keeping face after change
@@ -67,19 +63,17 @@ class FaceDetector():
         self.facenet_detecor.keep_all = self.keep_all
         return ret , prob
 
-
-
     def get_single_face(self,img, is_PIL_input, norm=True):
         # TODO this is a horrible function and flow. will later be upgrades with instace segmentaion?
         face_img, prob = self.facenet_detecor(img, return_prob=True)
         if is_img(face_img):
             if face_img.size()[0] > 1:  # two or more faces detected in the img crop
                 # faceClassifer.imshow(face_img[0:2])
-                face_img = self.crop_top_third_and_sides(img, is_PIL_input)
+                face_img = crop_top_third_and_sides(img, is_PIL_input)
                 if is_PIL_input:
                     face_img, prob = self.detect_single_face(face_img)  # this returns a single img of dim 3
                     if is_img(face_img) and norm:
-                        face_img = self.normalize_image(face_img)
+                        face_img = normalize_image(face_img)
                 else:
                     # face_img, prob = self.detect_single_face_inv_norm(face_img)
                     face_img, prob = None, 0
@@ -91,20 +85,21 @@ class FaceDetector():
                     #     plt.show()
             else:
                 face_img = face_img[0]  # current face_img shape is 1ximage size(dim=3), we only want the img itself
-                face_img = self.normalize_image(face_img) if norm else face_img
+                face_img = normalize_image(face_img) if norm else face_img
                 prob = prob[0]
         return face_img , prob
 
 
-    def filter_out_non_face_corps(self, recreate_data, save_images=False) -> None:
+    def filter_out_non_face_corps(self, recreate_data, save_images_path='') -> dict:
         """ Given a set of image crop filter out all images without the faces present
         and update the high conf face img member """
+
+        high_conf_face_imgs = defaultdict(list)
         if self.faces_data_path and not recreate_data:
             print("pickle path to images received, loading...")
-            self.high_conf_face_imgs = pickle.load(open(os.path.join(self.faces_data_path, 'images_with_crop.pkl'),'rb'))
+            high_conf_face_imgs = pickle.load(open(os.path.join(self.faces_data_path, 'images_with_crop.pkl'),'rb'))
         else:
-            norm = not save_images
-            print('here')
+            norm = not save_images_path
             face_crops = get_entries(filters={Crop.is_face == True,
                                               Crop.reviewed_one == True,
                                               Crop.is_vague == False,
@@ -112,6 +107,8 @@ class FaceDetector():
             crops_path = "/mnt/raid1/home/bar_cohen/"
             raw_imgs_dict = defaultdict(list)
             for i, crop in tqdm.tqdm(enumerate(face_crops), total=len(face_crops)):
+                if not i % 3 == 0:
+                    continue
                 # print(i ,'/',len(face_crops))
                 # fix for v, v_ issue
                 name = crop.im_name
@@ -126,13 +123,11 @@ class FaceDetector():
             print(f"Received a total of {given_num_of_images} images")
             counter = 0
             errors = 0
-            to_save ='/mnt/raid1/home/bar_cohen/FaceData/labeled_images/high_res/'
             for id in raw_imgs_dict.keys():
                 print('cur id is', id)
-                os.makedirs(os.path.join(to_save, str(ID_TO_NAME[id])), exist_ok=True)
                 for img,crop in raw_imgs_dict[id]:
                     print(f'{counter}/{given_num_of_images}')
-                    if save_images:
+                    if save_images_path:
                         pass
                         # plt.title(f'Original Crop, label: {ID_TO_NAME[id]} ,counter : {counter}')
                         # plt.imsave(os.path.join(to_save,str(ID_TO_NAME[id]),str(counter) + '_orig.jpg'),
@@ -140,86 +135,24 @@ class FaceDetector():
                     try:
                         ret, _ = self.get_single_face(img, is_PIL_input=True, norm=norm)
                         if is_img(ret):
-                            if save_images:
+                            if save_images_path:
+                                os.makedirs(os.path.join(save_images_path, str(ID_TO_NAME[id])), exist_ok=True)
                                 img_show = ret.permute(1, 2, 0).int().numpy().astype(np.uint8)
                                 plt.clf()
                                 # plt.title(f'Detected Face, label:  {ID_TO_NAME[id]}, counter : {counter}')
-                                plt.imsave(os.path.join(to_save, str(ID_TO_NAME[id]), f'{crop.im_name}'), img_show)
-                            self.high_conf_face_imgs[id].append((ret,crop))
+                                plt.imsave(os.path.join(save_images_path, str(ID_TO_NAME[id]), f'{crop.im_name}'), img_show)
+                            high_conf_face_imgs[id].append((ret,crop))
                         counter += 1
-                    except:
+                    except Exception as e:
+                        print(e)
                         errors += 1
 
-            given_num_of_images_final = sum([len(self.high_conf_face_imgs[i]) for i in raw_imgs_dict.keys()])
+            given_num_of_images_final = sum([len(high_conf_face_imgs[i]) for i in raw_imgs_dict.keys()])
             print(f'Post filter left with {given_num_of_images_final}')
             print(f'errors: {errors}')
+            pickle.dump(high_conf_face_imgs, open(os.path.join('/mnt/raid1/home/bar_cohen/FaceData/', 'images_with_crop.pkl'),'wb'))
+        return high_conf_face_imgs
 
-            pickle.dump(self.high_conf_face_imgs, open(os.path.join('/mnt/raid1/home/bar_cohen/FaceData/', 'images_with_crop.pkl'),'wb'))
-
-
-    def create_X_y_faces(self, save_images=False):
-        """
-        Iterate over high_conf_face_imgs and extract the X images and y labels
-        """
-        assert self.high_conf_face_imgs , "high conf face images must be non-empty"
-        X_train, X_val, X_test, y_train, y_val, y_test = [], [], [] , [], [], []
-        counter = 0
-        for k in self.high_conf_face_imgs.keys():
-            for img,crop in self.high_conf_face_imgs[k]:
-                date = crop.vid_name[4:8]
-                img_show = img.permute(1, 2, 0).int().numpy().astype(np.uint8)
-                img = self.normalize_image(img)
-                if date in ['0730','0808']:
-                    X_test.append(img)
-                    y_test.append(k)
-                    if save_images:
-                        cur_path = os.path.join('/mnt/raid1/home/bar_cohen/FaceData/DataSplit/','test',str(ID_TO_NAME[k]))
-                        os.makedirs(cur_path, exist_ok=True)
-                        plt.imsave(os.path.join(cur_path, f'{crop.im_name}'), img_show)
-
-                elif date in ['0802','0803']:
-                    X_val.append(img)
-                    y_val.append(k)
-                    if save_images:
-                        cur_path = os.path.join('/mnt/raid1/home/bar_cohen/FaceData/DataSplit/','val',str(ID_TO_NAME[k]))
-                        os.makedirs(cur_path, exist_ok=True)
-                        plt.imsave(os.path.join(cur_path, f'{crop.im_name}'), img_show)
-
-                elif date in ['0729', '0801','0804','0805']:
-                    X_train.append(img)
-                    y_train.append(k)
-                    if save_images:
-                        cur_path = os.path.join('/mnt/raid1/home/bar_cohen/FaceData/DataSplit/','train',str(ID_TO_NAME[k]))
-                        os.makedirs(cur_path, exist_ok=True)
-                        plt.imsave(os.path.join(cur_path, f'{crop.im_name}'), img_show)
-                else:
-                    print('WTF YOY OY OYO ')
-                counter += 1
-
-        return X_train ,y_train, X_val ,y_val, X_test, y_test
-
-    def smooth_sample_of_data(self, x:list, y_list, max_sample_tresh):
-
-        for k in self.high_conf_face_imgs.keys():
-            if len(self.high_conf_face_imgs[k]) > max_sample_tresh:
-                inds = np.random.randint(0, len(self.high_conf_face_imgs[k]), size=max_sample_tresh)
-                self.high_conf_face_imgs[k] = np.array(self.high_conf_face_imgs[k])[:,inds]
-
-
-    def build_samples_hist(self, le, dataloader, title:str=None):
-        """Create a Bar plot that represented the number of face samples from every id"""
-        samples_dict = defaultdict(int)
-        for _, batch_ids in dataloader:
-            for id in batch_ids:
-                samples_dict[ID_TO_NAME[le.inverse_transform([int(id)])[0]]] += 1
-        df = pd.DataFrame(samples_dict, columns=samples_dict.keys(),index=[0])
-        ax= sns.barplot(data=df)
-        ax.set_xticklabels(df.columns, rotation=45)
-        plt.title(title)
-        plt.show()
-
-def is_img(img):
-    return img is not None and img is not img.numel()
 
 def collect_faces_from_video(video_path:str) -> []:
     fd = FaceDetector(faces_data_path=None,thresholds=[0.97,0.97,0.97],keep_all=True)
@@ -238,38 +171,15 @@ def collect_faces_from_list_of_videos(list_of_videos:list):
         face_imgs.extend(collect_faces_from_video(video_path=video_path))
     return face_imgs
 
-
-def load_data(data_path):
-    le = pickle.load(open(os.path.join(data_path, 'le.pkl'), 'rb'))
-    dl_train = pickle.load(open(os.path.join(data_path, 'dl_train_1.pkl'), 'rb'))
-    dl_val = pickle.load(open(os.path.join(data_path, 'df_val_1.pkl'), 'rb'))
-    dl_test = pickle.load(open(os.path.join(data_path, 'df_test_1.pkl'), 'rb'))
-    return le, dl_train, dl_val, dl_test
-
-
-class FacesDataset(torch.utils.data.Dataset):
-    def __init__(self, images:[], labels:[]):
-        super(FacesDataset, self).__init__()
-        self.images = images
-        self.labels = labels
-
-    def __len__(self):
-        return len(self.labels)
-
-    def __getitem__(self, item):
-        label = self.labels[item]
-        image = self.images[item]
-        return image, label
-
 def main():
     """Simple test of FaceDetector"""
-    fd = FaceDetector(faces_data_path='/mnt/raid1/home/bar_cohen/FaceData/',device='cuda:0')
+    # fd = FaceDetector(faces_data_path='/mnt/raid1/home/bar_cohen/FaceData/',device='cuda:0')
     # fd.filter_out_non_face_corps(recreate_data=True, save_images=False)
     # from faceClassifer import load_data
-    le, dl_train, dl_val, dl_test =  load_data('/mnt/raid1/home/bar_cohen/FaceData/')
-    fd.build_samples_hist(le, dl_train, title='dataloader train')
-    fd.build_samples_hist(le, dl_val, title='dataloader val')
-    fd.build_samples_hist(le, dl_test, title='dataloader test')
+    # le, dl_train, dl_val, dl_test =  load_data('/mnt/raid1/home/bar_cohen/FaceData/')
+    # fd.build_samples_hist(le, dl_train, title='dataloader train')
+    # fd.build_samples_hist(le, dl_val, title='dataloader val')
+    # fd.build_samples_hist(le, dl_test, title='dataloader test')
     # X,y, _, _,_,_ = fd.create_X_y_faces() # only taining
     # print(len(X), len(y))
     # fd.build_samples_hist(fd.high_conf_face_imgs, 'Face Images Hist')
